Domain model User Role

+++ src/main/java/com/charlespereira/dscommerce/services/ProductService.java
@@ -59,6 +59,9 @@ public class ProductService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not found!");
+        }
         try {
             repository.deleteById(id);
         }
